+++ test/testdata/flavors_rst.py
@@ -68,3 +68,50 @@ def tables():
     | `Cell 1` | [Cell 2](http://example.com) link |
     | Cell 3 | **Cell 4** |
     """
+
+
+def footnote1():
+    """
+    Cite the relevant literature, e.g. [1]_.  You may also cite these
+    references in the notes section above.
+
+    .. [1] O. McNoleg, "The integration of GIS, remote sensing,
+       expert systems and adaptive co-kriging for environmental habitat
+       modelling of the Highland Haggis using object-oriented, fuzzy-logic
+       and neural-network techniques," Computers & Geosciences, vol. 22,
+       pp. 585-588, 1996.
+    """
+
+
+def footnote2():
+    """
+    Autonumbered footnotes are
+    possible, like using [#]_ and [#]_.
+
+    .. [#] This is the first one.
+    .. [#] This is the second one.
+
+    They may be assigned 'autonumber
+    labels' - for instance,
+    [#fourth]_ and [#third]_.
+
+    .. [#third] a.k.a. third_
+
+    .. [#fourth] a.k.a. fourth_
+    """
+
+
+def footnote3():
+    """
+    Auto-symbol footnotes are also
+    possible, like this: [*]_ and [*]_.
+
+    .. [*] This is the first one.
+    .. [*] This is the second one.
+    """
+
+
+def footnote4():
+    """
+    There is not footnote for this reference [#]_.
+    """